refactor: Piece 인터페이스 추가 및 Soldier 이동 구현
