+++ src/main/java/janggi/piece/DefaultUnitPosition.java
@@ -1,11 +1,9 @@
-package janggi.unit;
+package janggi.piece;
 
 import janggi.position.Position;
-import java.util.AbstractMap;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Map.Entry;
 import java.util.function.Supplier;
 
 public enum DefaultUnitPosition {
