--- src/main/java/janggi/unit/Team.java
@@ -1,14 +1,0 @@
-package janggi.unit;
-
-public enum Team {
-    HAN,
-    CHO,
-    ;
-
-    public Team getOppositie() {
-        if (this == HAN) {
-            return CHO;
-        }
-        return HAN;
-    }
-}
