--- src/test/java/janggi/unit/PositionTest.java
@@ -1,63 +1,0 @@
-package janggi.unit;
-
-import janggi.position.Position;
-import org.assertj.core.api.Assertions;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.ValueSource;
-
-class PositionTest {
-    @ValueSource(strings = {"0,-1,row가 장기판 범위를 벗어났습니다: -1", "-1,0,column이 장기판 범위를 벗어났습니다: -1", "9,0,column이 장기판 범위를 벗어났습니다: 9", "0,10,row가 장기판 범위를 벗어났습니다: 10"})
-    @ParameterizedTest
-    @DisplayName("좌표 범위를 넘으면 예외를 발생시킨다")
-    void test1(String str) {
-        // given
-        String[] split = str.split(",");
-        int x = Integer.parseInt(split[0]);
-        int y = Integer.parseInt(split[1]);
-        String exceptionMessage = split[2];
-
-        // when & then
-        Assertions.assertThatThrownBy(() -> new Position(x, y))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage(exceptionMessage);
-    }
-
-    @ValueSource(strings = {"0,2", "2,0", "2,9", "8,2"})
-    @ParameterizedTest
-    @DisplayName("수평 혹은 수직 직선에 점이 위치해 있으면 참을 반환한다")
-    void test2(String str) {
-        // given
-        String[] split = str.split(",");
-        int x = Integer.parseInt(split[0]);
-        int y = Integer.parseInt(split[1]);
-        Position position = new Position(x, y);
-
-        Position opposite = new Position(2, 2);
-
-        // when
-        boolean b = position.isStraightLine(opposite);
-
-        // then
-        Assertions.assertThat(b).isTrue();
-    }
-
-    @ValueSource(strings = {"3,3", "1,1", "3,1", "1,3"})
-    @ParameterizedTest
-    @DisplayName("수평 혹은 수직 직선에 점이 위치해있지 않으면 거짓을 반환한다")
-    void test3(String str) {
-        // given
-        String[] split = str.split(",");
-        int x = Integer.parseInt(split[0]);
-        int y = Integer.parseInt(split[1]);
-        Position position = new Position(x, y);
-
-        Position opposite = new Position(2, 2);
-
-        // when
-        boolean b = position.isStraightLine(opposite);
-
-        // then
-        Assertions.assertThat(b).isFalse();
-    }
-}
